Add `hooks.withTypes.test-d.tsx` type test file
