+++ test/typetests/hooks.tsx
@@ -18,10 +18,10 @@ import {
   useStore,
 } from '../../src/index'
 
-import type { AppDispatch, AppStore, RootState } from './counterApp'
+import type { AppDispatch, RootState } from './counterApp'
 import { incrementAsync } from './counterApp'
 
-import { exactType, expectExactType, expectType } from '../typeTestHelpers'
+import { expectExactType, expectType } from '../typeTestHelpers'
 
 function preTypedHooksSetup() {
   // Standard hooks setup
@@ -236,44 +236,4 @@ function testCreateHookFunctions() {
     ) => Selected
   >(createSelectorHook(Context))
   expectType<() => Store<RootState, RootAction>>(createStoreHook(Context))
-}
-
-function preTypedHooksSetupWithTypes() {
-  const useAppDispatch = useDispatch.withTypes<AppDispatch>()
-
-  const useAppSelector = useSelector.withTypes<RootState>()
-
-  const useAppStore = useStore.withTypes<AppStore>()
-
-  function CounterComponent() {
-    useAppSelector((state) => state.counter)
-
-    const dispatch = useAppDispatch()
-
-    expectExactType<AppDispatch>(dispatch)
-
-    const store = useAppStore()
-
-    expectExactType<AppStore>(store)
-
-    expectExactType<AppDispatch>(store.dispatch)
-
-    const state = store.getState()
-
-    expectExactType<RootState>(state)
-
-    expectExactType<number>(state.counter)
-
-    store.dispatch(incrementAsync(1))
-
-    exactType(store.dispatch, dispatch)
-
-    return (
-      <button
-        onClick={() => {
-          dispatch(incrementAsync(1))
-        }}
-      />
-    )
-  }
 }
